requirements.txt
gunicorn==20.1.0
azure-storage-blob==12.19.0
azure-core==1.34.0
python-dotenv==1.0.1
requests==2.32.3
PyMuPDF==1.25.5
python-docx==1.1.2
flask>=2.0 
Werkzeug==2.2.2
